Defect/Story Number:	OAS-4456 Description of Defect/Story: 	LLO RP - Modify Script for CT Customer Migration Who Did Code Review: 	Somenath Chakraborti Code Files Changes: 	/oasx/LASLinks - Replatforming/SETUP_LLO_RP_CUSTOMER_CONFIG.sql Impacted Areas/Regression Areas: 	Customer setup for LLO RP Customer How The Issue Was Solved: 	New input parameter added in the existing procedure "SETUP_LLO_RP_CUSTOMER_CONFIG" to determine whether the customer is newly created LLO customer or existing customer and program & test association data handling have been modified as per requirement. Hints To QA For Testing: 	Create new LLO customer using Legacy TAS module, and run the procedure to verify that the existing behavior of the script is intact for newly created customer. Run the procedure on existing LLO customer with old program and test data, and verify if it sets up the program, test associations and related configurations properly. It should be noted that old program end dates will be extended with value current date plus 10 years. Developer Comments: From QC. May be old but could provide some useful information 	1.	There are two inputs to the procedure - a) "in_customer_id" , b) "add_engrade_config", and c) “IS_NEW_CUSTOMER”. Provide the customer_id in the input "in_customer_id". If "add_engrade_config" input is “True" then the engrade related configurations will be given to the customer along with the LLO RP related configurations. If "add_engrade_config" input is "False" then only LLO RP related configurations will be given, the engrade related configurations will not be given. 2.	Run the procedure with new input parameter “IS_NEW_CUSTOMER” as TRUE if the customer is newly created LLO customer, and FALSE if the customer is an existing LLO customer with old program and test data. Please make sure to choose appropriate value for the flag as needed. 3.	The script should be run only once.

+++ oasx/LASLinks - Replatforming/SETUP_LLO_RP_CUSTOMER_CONFIG.sql
@@ -1,6 +1,7 @@
 create or replace
 PROCEDURE SETUP_LLO_RP_CUSTOMER_CONFIG(IN_CUSTOMER_ID     INTEGER,
-                                                         ADD_ENGRADE_CONFIG BOOLEAN) AS
+                                       ADD_ENGRADE_CONFIG BOOLEAN,
+                                       IS_NEW_CUSTOMER BOOLEAN) AS
 
   V_CHECK_CONFIG         NUMBER := 0;
   V_NEW_PRODUCT_ID       NUMBER := 0;
@@ -81,10 +82,16 @@ BEGIN
            IF V_CONFIGCOUNT > 0 THEN
              
             dbms_output.put_line('program start');
-            DELETE FROM PROGRAM PROG
-             WHERE PROG.PRODUCT_ID = ALL_PRODUCT_ID.PARENT_PRODUCT_ID
-               AND PROG.CUSTOMER_ID = IN_CUSTOMER_ID;
-          
+            IF (IS_NEW_CUSTOMER) THEN
+               DELETE FROM PROGRAM PROG
+               WHERE PROG.PRODUCT_ID = ALL_PRODUCT_ID.PARENT_PRODUCT_ID
+                 AND PROG.CUSTOMER_ID = IN_CUSTOMER_ID;
+            ELSE
+              UPDATE PROGRAM PROG SET PROG.ACTIVATION_STATUS = 'DE', PROG.PROGRAM_END_DATE = add_months(SYSDATE, 120)
+               WHERE PROG.PRODUCT_ID = ALL_PRODUCT_ID.PARENT_PRODUCT_ID
+                 AND PROG.CUSTOMER_ID = IN_CUSTOMER_ID;             
+            END IF;
+            
             SELECT CAST(IN_CUSTOMER_ID || '72' AS INT)
               INTO V_PROGRAM_ID
               FROM DUAL;
@@ -122,11 +129,19 @@ BEGIN
         /* Program end*/
         dbms_output.put_line('program end');
         /*ORG_NODE_TEST_CATALOG table start */
-        DELETE FROM ORG_NODE_TEST_CATALOG ONTC
-         WHERE ONTC.PRODUCT_ID = ALL_PRODUCT_ID.PRODUCT_ID
-           AND ONTC.CUSTOMER_ID = IN_CUSTOMER_ID;
       
         dbms_output.put_line('ONTC start');
+        IF (IS_NEW_CUSTOMER) THEN
+          DELETE FROM ORG_NODE_TEST_CATALOG ONTC
+           WHERE ONTC.PRODUCT_ID = ALL_PRODUCT_ID.PRODUCT_ID
+             AND ONTC.CUSTOMER_ID = IN_CUSTOMER_ID;
+        ELSE
+          UPDATE ORG_NODE_TEST_CATALOG ONTC
+             SET ONTC.ACTIVATION_STATUS = 'IN'
+           WHERE ONTC.PRODUCT_ID = ALL_PRODUCT_ID.PRODUCT_ID
+             AND ONTC.CUSTOMER_ID = IN_CUSTOMER_ID;
+        END IF;
+        
         SELECT COUNT(1)
           INTO V_CHECK_NEW_PRODUCT
           FROM ORG_NODE_TEST_CATALOG
@@ -252,9 +267,15 @@ BEGIN
            IF V_CONFIGCOUNT > 0 THEN
            
             dbms_output.put_line('Program start');
-            DELETE FROM PROGRAM PROG
-             WHERE PROG.PRODUCT_ID = ALL_PRODUCT_ID.PARENT_PRODUCT_ID
-               AND PROG.CUSTOMER_ID = IN_CUSTOMER_ID;
+            IF (IS_NEW_CUSTOMER) THEN
+               DELETE FROM PROGRAM PROG
+               WHERE PROG.PRODUCT_ID = ALL_PRODUCT_ID.PARENT_PRODUCT_ID
+                 AND PROG.CUSTOMER_ID = IN_CUSTOMER_ID;
+            ELSE
+              UPDATE PROGRAM PROG SET PROG.ACTIVATION_STATUS = 'DE', PROG.PROGRAM_END_DATE = add_months(SYSDATE, 120)
+               WHERE PROG.PRODUCT_ID = ALL_PRODUCT_ID.PARENT_PRODUCT_ID
+                 AND PROG.CUSTOMER_ID = IN_CUSTOMER_ID;             
+            END IF;
           
             SELECT CAST(IN_CUSTOMER_ID || '78' AS INT)
               INTO V_PROGRAM_ID
@@ -293,10 +314,18 @@ BEGIN
         /* Program end*/
       
         /*ORG_NODE_TEST_CATALOG table start */
-        DELETE FROM ORG_NODE_TEST_CATALOG ONTC
-         WHERE ONTC.PRODUCT_ID = ALL_PRODUCT_ID.PRODUCT_ID
-           AND ONTC.CUSTOMER_ID = IN_CUSTOMER_ID;
+ 
         dbms_output.put_line('ONTC start');
+        IF (IS_NEW_CUSTOMER) THEN
+          DELETE FROM ORG_NODE_TEST_CATALOG ONTC
+           WHERE ONTC.PRODUCT_ID = ALL_PRODUCT_ID.PRODUCT_ID
+             AND ONTC.CUSTOMER_ID = IN_CUSTOMER_ID;
+        ELSE
+          UPDATE ORG_NODE_TEST_CATALOG ONTC
+             SET ONTC.ACTIVATION_STATUS = 'IN'
+           WHERE ONTC.PRODUCT_ID = ALL_PRODUCT_ID.PRODUCT_ID
+             AND ONTC.CUSTOMER_ID = IN_CUSTOMER_ID;
+        END IF;
       
         SELECT COUNT(1)
           INTO V_CHECK_NEW_PRODUCT
